requirements.txt
flask
langchain
langchain-google-genai
python-dotenv
plyer
schedule